creating support for tenants in same db using prefixes

+++ src/Tenant/DatabaseConnection.php
@@ -15,6 +15,18 @@ use Hyn\MultiTenant\Models\Website;
 class DatabaseConnection
 {
     /**
+     * See the multi-tenant configuration file. Configuration set
+     * to use separate databases.
+     */
+    const TENANT_MODE_SEPARATE_DATABASE = 'database';
+
+    /**
+     * See the multi-tenant configuration file. Configuration set
+     * to use prefixed table in same database.
+     */
+    const TENANT_MODE_TABLE_PREFIX = 'prefix';
+
+    /**
      * @var Website
      */
     protected $website;
@@ -93,12 +105,22 @@ class DatabaseConnection
      * Generic configuration for tenant.
      *
      * @return array
+     * @throws TenantDatabaseException
+     * @throws \Laracasts\Presenter\Exceptions\PresenterException
      */
     protected function config()
     {
         $clone = Config::get(sprintf('database.connections.%s', static::systemConnectionName()));
-        $clone['password'] = md5(Config::get('app.key').$this->website->id);
-        $clone['username'] = $clone['database'] = sprintf('%d-%s', $this->website->id, $this->website->present()->identifier);
+
+        if (Config::get('multi-tenant.db.tenant-division-mode') == static::TENANT_MODE_SEPARATE_DATABASE) {
+            $clone['password'] = md5(Config::get('app.key') . $this->website->id);
+            $clone['username'] = $clone['database'] = sprintf('%d-%s', $this->website->id,
+                $this->website->present()->identifier);
+        } elseif (Config::get('multi-tenant.db.tenant-division-mode') == static::TENANT_MODE_TABLE_PREFIX) {
+            $clone['prefix'] = sprintf('t%d_', $this->website->id);
+        } else {
+            throw new TenantDatabaseException("Unknown database division mode configured in the multi-tenant configuration file.");
+        }
 
         return $clone;
     }
@@ -116,13 +138,17 @@ class DatabaseConnection
      */
     public function create()
     {
+        if (Config::get('multi-tenant.db.tenant-division-mode') != static::TENANT_MODE_SEPARATE_DATABASE) {
+            return null;
+        }
+
         $clone = $this->config();
 
         return DB::connection(static::systemConnectionName())->transaction(function () use ($clone) {
-            if (! DB::statement("create database if not exists `{$clone['database']}`")) {
+            if (!DB::statement("create database if not exists `{$clone['database']}`")) {
                 throw new TenantDatabaseException("Could not create database {$clone['database']}");
             }
-            if (! DB::statement("grant all on `{$clone['database']}`.* to `{$clone['username']}`@'localhost' identified by '{$clone['password']}'")) {
+            if (!DB::statement("grant all on `{$clone['database']}`.* to `{$clone['username']}`@'localhost' identified by '{$clone['password']}'")) {
                 throw new TenantDatabaseException("Could not create or grant privileges to user {$clone['username']} for {$clone['database']}");
             }
 
@@ -137,16 +163,20 @@ class DatabaseConnection
      */
     public function delete()
     {
+        if (Config::get('multi-tenant.db.tenant-division-mode') != static::TENANT_MODE_SEPARATE_DATABASE) {
+            return null;
+        }
+        
         $clone = $this->config();
 
         return DB::connection(static::systemConnectionName())->transaction(function () use ($clone) {
-            if (! DB::statement("revoke all on `{$clone['database']}`.* from `{$clone['username']}`@'localhost'")) {
+            if (!DB::statement("revoke all on `{$clone['database']}`.* from `{$clone['username']}`@'localhost'")) {
                 throw new TenantDatabaseException("Could not revoke privileges to user {$clone['username']} for {$clone['database']}");
             }
-            if (! DB::statement("drop database `{$clone['database']}`")) {
+            if (!DB::statement("drop database `{$clone['database']}`")) {
                 throw new TenantDatabaseException("Could not drop database {$clone['database']}");
             }
-            if (! DB::statement("drop user `{$clone['username']}`@'localhost'")) {
+            if (!DB::statement("drop user `{$clone['username']}`@'localhost'")) {
                 throw new TenantDatabaseException("Could not drop user {$clone['username']}");
             }
